pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gpcr-annotation-tools"
version = "0.1.3"
description = "Human-in-the-loop curation suite for GPCR structure annotations"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "GPCR Annotation Team" },
]
keywords = ["gpcr", "pdb", "annotation", "structural-biology", "curation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "rich>=13.0",
    "tqdm>=4.60",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-requests>=2.31",
    "types-tqdm>=4.60",
    "gpcr-annotation-tools[annotate,papers]",
]
annotate = [
    "google-genai",
]
papers = [
    "PyMuPDF",
    "lxml",
]

[project.scripts]
gpcr-tools = "gpcr_tools.__main__:cli"
gpcr-csv-generator = "gpcr_tools.csv_generator.app:cli_entry"

[tool.hatch.build.targets.wheel]
packages = ["src/gpcr_tools"]

# ── Ruff (Linter + Formatter) ──────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator (readability preference)
]

[tool.ruff.lint.isort]
known-first-party = ["gpcr_tools"]

# ── Mypy ────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false          # gradual adoption
check_untyped_defs = true
ignore_missing_imports = false

# ── Pytest ──────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
